cv/scripts/requirements.txt
jinja2>=3.1
pyyaml>=6.0
weasyprint>=60.0
